+++ tests/valid_seq/spurious_match.m
@@ -28,6 +28,8 @@ main(!IO) :-
     :- module spurious_match.spurious_match_helper_1.
     :- interface.
 
+    :- import_module list.
+
     :- pred parse_tokens(string::in, list(int)::in, list(int)::out) is det.
 
     :- implementation.